reorganize filetree for components
