+++ src/App.js
@@ -7,10 +7,10 @@ import React, { useState, Suspense } from 'react';
 /* COMPONENTS */ 
 import ImportedMode from './components/peach/ImportedMode';
 import Footer from './components/Footer';
-import SchemaDisplayContainer from './components/SchemaDisplayContainer';
-import WrittenResponseDisplay from './components/WrittenResponseDisplay';
-import QueryContainer from './components/QueryContainer';
-import VariableInput from './components/VariableInput';
+import SchemaDisplayContainer from './components/SchemaDisplay/SchemaDisplayContainer';
+import WrittenResponseDisplay from './components/ResponseDisplay/WrittenResponseDisplay';
+import QueryContainer from './components/QueryContainer/QueryContainer';
+import VariableInput from './components/VariableInput/VariableInput';
 import ErrorBoundary from './components/ErrorBoundary';
 
 /* STYLES */
