+++ src/components/QueryEditor/EditorDisplay.js
