+++ src/components/ResponseDisplay/WrittenResponseDisplay.js
@@ -4,7 +4,7 @@ stateless component renders query response data passed from App.js
 
 import React from 'react';
 import Container from 'react-bootstrap/Container';
-import '../styles/styles.css';
+import '../../styles/styles.css';
 
 const WrittenResponseDisplay = ({response}) => {
