+++ src/components/SchemaDisplay/SchemaDisplay.js
