+++ src/components/SchemaDownload/InputGqlSchema.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import Container from 'react-bootstrap/Container';
 import Button from 'react-bootstrap/Button';
 import Modal from './Modal';
-import gqlEndpoint from '../relay/gqlEndpoint';
+import gqlEndpoint from '../../relay/gqlEndpoint';
 
 const InputGqlSchema = () => {
   const [showModal, setShowModal] = useState(false);
